feat: 유저 삭제 시 존재여부에 따라 응답 분기처리

+++ src/main/java/com/swmStrong/demo/domain/user/service/UserServiceImpl.java
@@ -102,8 +102,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(String userId) {
-        userRepository.deleteById(userId);
+        if (!userRepository.existsById(userId)) {
+            throw new ApiException(ErrorCode.USER_NOT_FOUND);
+        }
 
+        userRepository.deleteById(userId);
     }
 
     private String getKey(String requestIP) {
